// apps/web-services/src/models/paginatedApi.model.ts
/*               "Copyright 2020 Infosys Ltd.
               Use of this source code is governed by GPL v3 license that can be found in the LICENSE file or at https://opensource.org/licenses/GPL-3.0
               This program is free software: you can redistribute it and/or modify it under the terms of the GNU General Public License version 3" */
import { IContent } from './content.model'

export interface IPaginatedApiResponse {
  contents: IContent[]
  hasMore?: boolean
  pageState?: string
  totalHits?: number
}